pycommons/requirements.txt
/tmp/apa102-pi
/tmp/raspberry-gpio

pyaudio
pyyaml
gpiozero
#rpi.gpio